### 2016-08-19 #53 @whedon
[REVIEW]: CAZy-parser a way to extract information from the Carbohydrate-Active enZYmes Database
**Submitting author:** @rodrigovrgs (<a href="http://orcid.org/0000-0001-5267-3002">Rodrigo Honorato</a>)
**Repository:** <a href="https://github.com/rodrigovrgs/cazy-parser" target ="_blank">https://github.com/rodrigovrgs/cazy-parser</a>
**Version:** v1.0
