### 2016-08-19 #53 @kyleniemeyer
That's strange—the "*Submitting author*" looks correct
