### 2016-08-19 2d3a971
Fixing formatting so paper will compile :-)
