+++ paper/paper.md
@@ -1,18 +1,18 @@
 ---
- title: 'CAZy-parser a way to extract information from the Carbohydrate-Active enZYmes Database'
- tags:
-   - datamining
-   - cazy
- authors:
+title: 'CAZy-parser a way to extract information from the Carbohydrate-Active enZYmes Database'
+tags:
+ - datamining
+ - cazy
+authors:
   - name: Rodrigo V Honorato
   - orcid: 0000-0001-5267-3002
   - affiliation: Brazilian Center for Research in Energy and Materials (CNPEM) / Brazilian Biosciences National Laboratory
- date: 18 August 2016
- bibliography: paper.bib
- ---
+date: 18 August 2016
+bibliography: paper.bib
+---
 
- # Summary
+# Summary
 
 The Carbohydrate-Active enZYmes Database (CAZy) provides access to a sequence based classification of enzyme that are responsible for the assembly, modification and breakdown of oligo and polysaccharides [@Lombard2013]. This database has been online for eighteen years providing relevant genomic, structural and biochemical data on carbohydrate-active enzymes, such as glycoside hydrolases, glycosyl transferases, polysaccharide lyases, carbohydrate esterases and similar enzymes with auxiliary activities. The database is organized and presented to the user as a series of highly annotated HTML tables. In this paper a script to extract information from the CAZy according to user needs is presented. To this date, there is no way to obtain specific subsets from the database. Therefore the scripts presented in this paper will enable researchers to carry out a wide range of complex scientific tasks such as; phylogenetic analyses, multiple sequence alignment and cross-family comparisons.
 
- # References
+# References
