### 2016-08-19 #53 @arfon
> That's strange—the "Submitting author" looks correct

Yeah but the submitting author also just enters their GitHub handle manually which it looks like @rodrigovrgs did incorrectly.
