### 2016-08-22 fed5a3b
removing bugs
