### 2016-08-22 #53 @rodrigovrgs
> That's strange—the "Submitting author" looks correct
>> Yeah but the submitting author also just enters their GitHub handle manually which it looks like @rodrigovrgs did incorrectly.

Sorry for that, I'm the @rodrigovrgs you are looking for. @arfon, is there a way for me to fix handle?
