### 2016-08-22 #53 @arfon
@rodrigovrgs - I've updated this thread with the correct GitHub handle. You can just change your GitHub username on JOSS here: http://joss.theoj.org/profile
