### 2016-09-01 #53 @pjotrp
I am not in a position to review this paper, but I would like to note this software consists of about 400 lines of code. I would like to ask the author to confirm *here* that he thinks this software is a valuable contribution to research. I want to point out that once this work has been published it will be visible forever. Adding an example of having used this tool in research would be *strongly* recommended.
