### 2016-09-10 rodrigovrgs/cazy-parser#2 @jkahn
organizing as a setup.py compatible project for pypi etc
Here from [the JOSS review thread](https://github.com/openjournals/joss-reviews/issues/53):

This project is not organized in the canonical ways a distributable piece of code should be in the standard Python universe.  I recommend reviewing https://packaging.python.org/ and consider distributing these as entry_points console scripts.
Organizing and distributing your code in standard ways makes it more likely that other OSS users can take advantage of your brilliance.
