### 2016-10-14 #53 @luizirber
The software is useful and I think it should be accepted, but some points need to be addressed:

- Installation instruction can be clearer: it only talks about dependencies,
but doesn't say that the repository needs to be downloaded or where to get the
released version. Can you add a setup.py and publish the package to
PyPI? `pip` is the standard way of installing Python packages, this also
increases the reach of your software. I see @jkhan already added an issue for this: https://github.com/rodrigovrgs/cazy-parser/issues/2

- Functionality documentation: there is no exposed API besides the two scripts,
for using from a shell. One of them doesn't have options, and the other one uses `argparse` for options.
Even if there are no options `create-cazy-db.py` could have a `--help` option to at least explain what the command does.

- Automated tests: not available.

- Community guidelines: Kind of. There is an email for contact, but no guidelines
on how to collaborate. A CONTRIBUTING.md file in the repo goes a long way on
addressing this.
