### 2016-10-14 #53 @luizirber
Sorry for the delay @pjotrp. @rodrigovrgs, I can help with more details on how to address these points if you want, should I open some issues in the `cazy-parser` repo?
