### 2016-10-20 #53 @rodrigovrgs
@pjotrp @luizirber I'll address the issues as soon as I can, thank you very much!
