### 2016-10-20 rvhonorato/cazy-parser#2 @rvhonorato
thank you @jkahn I'll look into it.
