### 2016-10-21 96b24ec
reviwed code according to joss feedback
