+++ cazy_parser/create_cazy_db.py
@@ -1,0 +1,225 @@
+#!/usr/bin/env python
+#==============================================================================#
+# Copyright (C) 2016  Rodrigo Honorato
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>
+#==============================================================================#
+
+#==============================================================================#
+# create a parsed database exploting CAZY html structure
+import os, sys, urllib, re, string, time, string, argparse
+from bs4 import BeautifulSoup
+
+def logo():
+	print '''
+		  ___   __   ____  _  _     ____   __   ____  ____  ____  ____
+		 / __) / _\ (__  )( \/ )___(  _ \ / _\ (  _ \/ ___)(  __)(  _ \\
+		( (__ /    \ / _/  )  /(___)) __//    \ )   /\___ \ ) _)  )   /
+		 \___)\_/\_/(____)(__/     (__)  \_/\_/(__\_)(____/(____)(__\_)
+
+		A simple way to retrieve fasta sequences from CAZy Database (:
+
+			This is the database creator script.
+
+'''
+
+def main():
+	logo()
+	parser = argparse.ArgumentParser(description='Generate a comma separated table with information gathered from the CAZy database; internet connection is required.')
+	args = parser.parse_args()
+
+	#==============================================================================#
+	# Species part
+	#==============================================================================#
+	print '>> Gathering species codes for species with full genomes'
+	# a = archea // b = bacteria // e = eukaryota // v = virus
+	species_domain_list = ['a', 'b', 'e', 'v']
+	species_dic = {}
+	for initial in string.uppercase:
+		for domain in species_domain_list:
+			link = 'http://www.cazy.org/%s%s.html' % (domain, initial)
+			f = urllib.urlopen(link)
+			species_list_hp = f.read()
+			# parse webpage
+			index_list = re.findall('"http://www.cazy.org/(b\d.*).html" class="nav">(.*)</a>', species_list_hp)
+			for entry in index_list:
+				index, species = entry
+				try:
+					species_dic[species].append(index)
+				except:
+					species_dic[species] = [index]
+
+	# Double check to see which of the species codes are valid
+	for species in species_dic:
+		entry_list = species_dic[species]
+		if len(entry_list) > 1:
+			# More than one entry for this species
+			#  > This is (likely) a duplicate
+			#  > Use the higher number, should be the newer page
+			newer_entry = max([int(i.split('b')[-1]) for i in entry_list])
+			selected_entry = 'b%i' % newer_entry
+
+			species_dic[species] = selected_entry
+		else:
+			species_dic[species] = species_dic[species][0]
+
+	#==============================================================================#
+	# Enzyme class part
+	#==============================================================================#
+
+	enzyme_classes = ['Glycoside-Hydrolases',
+		'GlycosylTransferases',
+		'Polysaccharide-Lyases',
+		'Carbohydrate-Esterases',
+		'Auxiliary-Activities']
+
+	db_dic = {}
[…]
