+++ cazy_parser/extract_cazy_ids.py
@@ -1,0 +1,158 @@
+#!/usr/bin/env python
+#==============================================================================#
+# Copyright (C) 2016  Rodrigo Honorato
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>
+#==============================================================================#
+
+#==============================================================================#
+# Modules
+import os, sys, itertools, urllib, argparse
+#==============================================================================#
+
+def logo():
+	print '''
+		  ___   __   ____  _  _     ____   __   ____  ____  ____  ____
+		 / __) / _\ (__  )( \/ )___(  _ \ / _\ (  _ \/ ___)(  __)(  _ \\
+		( (__ /    \ / _/  )  /(___)) __//    \ )   /\___ \ ) _)  )   /
+		 \___)\_/\_/(____)(__/     (__)  \_/\_/(__\_)(____/(____)(__\_)
+
+		A simple way to retrieve fasta sequences from CAZy Database (:
+
+			This is the accession code retrieval script.
+
+'''
+
+def main(argv=sys.argv[1:]):
+	logo()
+	#==============================================================================#
+	# Options
+	#==============================================================================#
+	parser = argparse.ArgumentParser(description='Select accession codes for a given protein family. Optional: Select subfamilies and/or characterized enzymes')
+
+	parser.add_argument('--db', action = "store", dest = 'db_file',
+		help = 'Database file generated by cazy-parser')
+
+	parser.add_argument('--family', action = "store", dest = 'target_family',
+		help = 'Family to be searched ex. GH5')
+
+	parser.add_argument('--subfamily', action = "store_true", default = False,
+		help = '(Optional) Create a file with accession codes for each subfamily')
+
+	parser.add_argument('--characterized', action = "store_true", default = False,
+		help = '(Optional) Create a file with accession codes only for characterized enzymes')
+
+	results = parser.parse_args()
+
+	check = False
+	if results.db_file == None:
+		print '\n>> [ERROR] Missing database file\n'
+		check = True
+
+	if results.target_family == None:
+		print '\n>> [ERROR] Missing target family\n'
+		check = True
+
+	if check == True:
+		parser.print_help()
+		sys.exit(0)
+
+	#==============================================================================#
+	# Input
+	#==============================================================================#
+	db_f = open(results.db_file).readlines()
+	header = db_f[0].split('\t')
+
+	print '>> Loading %s' % results.db_file
+	db = {}
+	for i, l in enumerate(db_f[1:]):
+		db[i] = {}
+		data = l.split('\t')
+		# init dictionary
+		for c, idx in enumerate(header):
+			try:
+				db[i][idx] = data[c]
+			except IndexError:
+				db[i][idx] = ''
+
+	#==============================================================================#
+	# Select all family
+	#==============================================================================#
+	if bool(results.target_family):
+
+		selection_list = []
+		for e in db:
+			if db[e]['family'] == results.target_family:
+				selection_list.append(db[e]['genbank'])
+
+		print '>> Retrieving all %i accession codes for Family %s' % (len(selection_list), results.target_family)
+
[…]
