+++ cazy_parser.egg-info/top_level.txt
@@ -1,0 +1,2 @@
+cazy_parser
+paper
