+++ CONTRIBUTE.md
@@ -1,0 +1,19 @@
+## cazy-parser
+*A way to extract specific information from the Carbohydrate-Active enZYmes.*
+
+# How to contribute?
+
+There are still a few features that could be implemented, such as:
+
+* Organism specific selection
+* Retrieve three dimensional structures for each family
+
+and specially
+
+* **Retrieve fasta sequences from NCBIs servers**
+
+___
+
+Feel free to contact me with **suggestions**, **bugs reports** or if you need any **assistance** running the softare.
+
+*rvhonorato at gmail.com*
