+++ README.rst
@@ -1,0 +1,19 @@
+cazy-parser
+============
+
+The `Carbohydrate-Active enZYmes Database (CAZy) <https://www.cazy.org>`_ provides access to a sequence based classification of enzyme that are responsible for the assembly, modification and breakdown of oligo and polysaccharides.
+
+This database has been online for eighteen years providing relevant genomic, structural and biochemical data on carbohydrate-active enzymes, such asglycoside hydrolases, glycosyl transferases, polysaccharide lyases, carbohydrateesterases and similar enzymes with auxiliary activities. The database isorganized and presented to the user as a series of highly annotated HTML tables.
+
+This script provides a way to extract information from the database according to user need.
+
+Installation
+============
+::
+
+    pip install cazy-parser
+
+Documentation
+=============
+
+Please refer to the `project page <https://github.com/rodrigovrgs/cazy-parser/blob/master/README.md>`_  for usage and more information
