### 2016-10-21 c1d3795
installation tweaks
