+++ cazy_parser.egg-info/top_level.txt
@@ -1,2 +1,1 @@
 cazy_parser
-paper
