### 2016-10-21 #53 @rodrigovrgs
@luizirber thanks for your review, I've made some changes to the code and the documentation to fit your feedback and make the code compliant to joss.

> Installation instruction can be clearer: it only talks about dependencies,
but doesn't say that the repository needs to be downloaded or where to get the
released version. Can you add a setup.py and publish the package to
PyPI? pip is the standard way of installing Python packages, this also
increases the reach of your software. I see @jkhan already added an issue for this: rodrigovrgs/cazy-parser#2

I've added the functionality and the issue has been closed.

> Functionality documentation: there is no exposed API besides the two scripts,
for using from a shell. One of them doesn't have options, and the other one uses argparse for options.
Even if there are no options create-cazy-db.py could have a --help option to at least explain what the command does.

The recently implement setup.py installation method makes it so that the scripts can be run from the command line. Also, --help arg has been implemented in the code as well as a means to display the user what argument is missing if that is the case.

> Automated tests: not available.

I'm not quite sure how to produce this, could you please give me one example?

> Community guidelines: Kind of. There is an email for contact, but no guidelines
on how to collaborate. A CONTRIBUTING.md file in the repo goes a long way on
addressing this.

I've added a CONTRIBUTING.md stating which features would be a great addiditon to the cazy-parser.
