### 2016-11-03 #53 @rodrigovrgs
@luizirber when could you take another look at this?
