### 2016-11-07 #53 @luizirber
Great work @rodrigovrgs!

About the tests: it is one of the journal requirements, but it is not so easy to test your scripts (because it depends on a big DB generated with one of them). You could set up a subset of the DB and execute the `extract_cazy_ids.py` on it, and check the outputs. Another alternative would be refactoring your scripts into functions to make it easier to test, but I think this gets out of the scope of the submission.

@pjotrp, what would be the recommendation in this case?
