### 2016-11-08 #53 @arfon
> @arfon I think we can accept this paper. @rodrigovrgs can look into improving the tests - but I think we are good to go.

👍 ok thanks.
