### 2016-12-12 #53 @arfon
> @rodrigovrgs - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.

Bump @rodrigovrgs - we're ready to accept this submission but need you to complete this last step before we can proceed.
