### 2016-12-13 5d7e7b3
updating doi badge with relase v1.0
