### 2016-12-13 #53 @rodrigovrgs
I've created an archive with the reviewed software [at Zenodo](https://zenodo.org/record/200893), updated the doi and the master's branch *README.md*.

[![DOI](https://zenodo.org/badge/65995178.svg)](https://zenodo.org/badge/latestdoi/65995178)

Is that everything?
