### 2016-12-14 5105a2e
Update paper.md
